Practica 5

+++ app/Http/Controllers/ProductoController.php
@@ -12,6 +12,10 @@ class ProductoController extends Controller
     }
 
     public function store(Request $request) {
+        if (!$request->user()->tokenCan('crear')) {
+            return response()->json(['mensaje' => 'Acción no autorizada. El token no tiene la habilidad: crear'], 403);
+        }
+
         $data = $request->validate([
             'nombre' => 'required|string|max:150',
             'precio' => 'required|numeric|min:0',
@@ -27,11 +31,19 @@ class ProductoController extends Controller
     }
 
     public function update(Request $request, Producto $producto) {
+        if (!$request->user()->tokenCan('editar')) {
+            return response()->json(['mensaje' => 'Acción no autorizada. El token no tiene la habilidad: editar'], 403);
+        }
+
         $producto->update($request->all());
         return response()->json($producto, 200);
     }
 
-    public function destroy(Producto $producto) {
+    public function destroy(Request $request, Producto $producto) {
+        if (!$request->user()->tokenCan('eliminar')) {
+            return response()->json(['mensaje' => 'Acción no autorizada. El token no tiene la habilidad: eliminar'], 403);
+        }
+
         $producto->delete();
         return response()->json(['mensaje' => 'Eliminado'], 200);
     }
